refactor(core): split debug tools into one-file-per-tool structure
